+++ client/src/components/QuestionPanel.vue
@@ -11,6 +11,7 @@
         <span style="font-size: 14px;">
           <UserName
             :name="userName"
+            :uid="question.uid"
           />
         </span>
         </v-card-title>
